Get the unit tests to run, and fix a few of the issues that they revealed.  Not all issues are fixed though.
Currently there is no basic build, so those tests are commented out.

Also, make the public api export the MathQuill object, and make the
webpack entry point src/index.js.  That imports the MathQuill object,
and sets it as the global MathQuill object.  This makes it possible to
use MathQuill as a module without the global MathQuill variable being
set.