Only add parenthesis around the numerator and denominator of a fraction in text if the fraction contains a binary operator (that is not a unary operator), a fraction, a summation notation, an integral, a space, a comma, a colon, or a semicolon.  Also remove the extra parentheses around the entire fraction, and instead put spaces on both ends unless the fraction is immediately preceded by a unary operator.
For superscripts and subcripts parentheses are only added when needed
also (in much the same as the fraction cases), and a space is added at
the end if there is a digit to the right.

Change the spaceBehavesLikeTab option to only work when inside of a math
block.  So space is still a space at depth 0 (outside of a math block).