Rework keyboard input and update options
This reworks MathQuill's keyboard event handling to not use the
deprecated "which" and "keyCode" KeyboardEvent attributes.  The
deprecated "keypress" event usage is also removed.  One thing gained
from this rewrite is that spaceBehavesLikeTab now works on touch screen
devices that utilize a virtual keyboard.  That was a severe defficiency
of the previous code.

Note for future work:  The saneKeyboardEvents.test.js unit tests need to
be reworked.  Currently they are a complete sham.  Sending a bunch of
made up events does not properly emulate browser behavior.  It merely
assumes certain behavior, and checks that things work with that assumed
behavior.  If some browser deviates from that behavior then things can
fail for that browser, and the unit tests won't tell you anything. The
tests will even succeed when run in that browser.

New methods have been added to the Options object to add and remove
autoCommands and autoOperatorNames.  A getter has been added to API
objects (which all inherit from the AbstractMathQuill object where this
is implemented) to facilitate using these methods.  To protect the
Options object instance variables these are all made private (via the #
prefix which is now supported on all major browsers).  So the only
access to these things is via the getters and setters of the Options
object.

The capability of adding and removing autoCommand's and
autoOperatorName's has been added to the input-test.html page.