When a fraction is started to the right of an operator name, do not put the operator name into the numerator.
Add a space after operator names in the text output unless it is
followed by a Bracket or Fraction (which adds parentheses around the
text after an operator name).

Make the supSubsRequireOperand better by additionally not allowing a
SupSub on a BinaryOperator, comma, semicolon, or colon.  Also make it so
that this takes effect even when there is a selection.  Previously this
made it possible to still create a SupSub without an operator.