Fix an issue that is caused by an event being re-dispatched to the wrong element.  The event is instead dispatched to the same element it originally occured on and this causes an infinite recursion.
Fix another issue when the autoSubscriptNumerals option is used.  If
`x3` is typed (which shows as x_3), then the down arrow is pressed to go
down into the subscript, and then 5 is typed, then the 5 is output
twice.  This is due to the need of a return statement in the MathBlock
writeMethodMixin write method.

If the rootsAreExponents option is enabled (the default for WeBWork),
and a superscript is on the nthroot, then add parentheses around it in
the text output.  For example if "root3 4 ^6" is typed, then this will
display as the 6th power of the cubed root of 4 (assuming the root
autoCommand is enabled -- also a WeBWorK default).  This makes the text
output in that case "((4)^(1/3))^6".  The current text output is
"(4)^(1/3)^6" which is not mathematically equivalent.  The latex output
is correct.

If something to the left of as SupSub (exponent or subscript) is deleted
and the new thing to the left is no longer valid for a SupSub to be on
then bring its contents out and remove the SupSub.  For example,
parentheses are automatically added when a SupSub is placed onto a
fraction.  If the parentheses are deleted, then this happens. Also if
the supSubsRequireOperand option is enabled (another WeBWorK default)
and the base is deleted, then this happens also.  Furthermore, if the
new thing on the left is another SubSub, then this is merged with that
one.

Also make it so that an nthroot with an index of 2 is submitted as a
square root.

Fix the latex command `\ `.  This was a matter of checking for the latex
string begin undefined, when it should have been a check if it was empty
or undefined.