Make trig functions and logarithms commands instead of a operator names.
These are now a new MathFunction object.

Now typing "\sin " (or any other trig function) creates a function and
is displayed as "sin( )".  If "sin" is an autoCommand then typing "sin"
without the backslash does this (see more on this below).  Note that
after typing "sin" the cursor will be before the parenthesis, where you
can add superscripts and subscripts.  However, nothing else is allowed
to be inserted before the parentheses.  Typing anything else before the
parentheses will automatically move into the parentheses and insert
there.  So typing "sinx" will give "sin(x)".  Typing "sin^2 x" will give
"sin^2(x)" (assuming spaceBehavesLikeTab is true).  MathFunction objects
are also similar to Bracket objects (parentheses, brackets, and braces),
but are only allowed to match and ending parenthesis.  This means if
"x+3" already exists to the right of the cursor, then typing "sin" will
give "sin(x+3)" (the content to the right is adopted).

It is also possible to extend or shorten a math functions name.  Thus sin
can become sinh by typing "h" at the end of "sin", and sinh can become
sin again by typing a "Backspace" at the end of "sinh".

There is a special case for the text output of logarithms with a base.
Instead of outputing "log_2(x)", by default that will now be
"logb(2,x)".  Note "log_2^3(x)" will be "(logb(2,x))^3".  However, if
the new option logsChangeBase is true (it is false by default), then the
text output will be "log(x)/log(2)" or "(log(x)/log(2))^3" with an
expontent.  Note that in any case the above is ignored for base 10.  For
base 10 the output is always "log10(x)".  Also note that the intention
is to set this option to true by default for the WeBWorK mqeditor.

Unit tests for this new MathFunction object have been added.  These unit
tests include tests for text output.  Note these are the first such
tests.  As important as text output is for WeBWorK, much more of this is
needed for the rest of MathQuill's objects.

With this change, the trig functions are no longer operator names.
Instead they are LaTeX commands.  This means that to get the previous
behavior where typing "sin" (for instance) without a preceding backslash
creates the "sin" function, "sin" must be added as an autoCommand.