A different approach to making degC, degF, and degK work for webwork.
This removes deg from the autoOperatorNames as did #6.  However, this
also changes the text output for the latex command \deg (and \degree) to
deg (it was degree).  Furthermore, if C, F, or K is typed after that
there is no space added between deg and that letter.  There is also a
small tweak so that a space is added before deg in the text output if
anything is to the left of it.  So if deg is added to the autoCommands
option and one types "32degF" the text output (submitted for the answer
in webwork) will be "32 degF", and "32°F" will be displayed in the
visible MathQuill input.