Fix some long standing issues with SupSubs with the supSubsRequireOperand option.
This ensures that if the supSubsRequireOperand option is true, a SupSub
cannot be left dangling where is should not be allowed to be with that
option.  There are numerous ways that this can happen currently.

For example, if x^2 is typed and then the arrow keys are used to move
the cursor just after x and a binary operator or space is typed, then
the superscript is still there.

Similarly (although different handling is needed in the code for this)
if x^2 is typed and the arrow keys are used to again move the cursor
just after x and a left parenthesis is typed, then the superscript is
left floating inside the parentheses that are created.

To fix this siblingDeleted and siblingCreated handlers need to be added
to "flatten" the SupSub when things are typed in between a SupSub and
the object it is intended to be on that are no longer valid for a SupSub
to be on.  In addition, the existing method I previously added that
partially dealt with this needed to be updated by always passing the
cursor in to the SupSub contactWeld method.

This also fixes related issues with autoOperatorName spacing.  The
siblingDeleted and siblingCreated handler monkey patching that was used
was not reliable.  So instead the handlers used to fix the above issue
are also used to reliably determine if the autoOperatorName spacing
needs to be adopted to the right of a SupSub on the operator name.