Add prettier and reformat the code.
I am not entirely happy with what prettier does with formatting, and
wish there were more options for some things.  However, it is nice to be
able to let it take care of formatting the code, and not have to be as
mindful of it yourself.

The eslint rules also were updated to use those from the
`@stylistic/eslint-plugin`.  The corresponding rules of `eslint` are
deprecated.