New inequality: != is ≠
Typing != turns into ≠ just as typing <= turns into ≤ or typing >= turns
into ≥.

The FactorialOrNEQ class handles this.  It derives from the Inequality
class, and makes ! a unary operator (which is oddly a strict inequality)
and typing = after it turns it into \\ne or ≠ (which is a binary
operator and a non strict inequality).