Typing space after a comma is always a space.
If the `spaceBehavesLikeTab` option is true (as it is by default for
webwork2 and PG), then typing a space after a comma is still a space and
does not behave like tab there.  Typing a space after a comma is a good
habit that should not be punished.

This fixes issue #22.