Allow one period in a Fraction numerator or denominator or in a SupSub.
This means that if one types `5^2.3`, then that will also be the text
sent to WeBWorK.  No additional parentheses will be added around the
exponent.

Also `5.2/2.3` will be sent as is with no additional parentheses.

However, `5^2.3.4` will be sent as `5^(2.3.4)` and `5.2/2.3.4` will be
sent as `5.2/(2.3.4)`.

This is to address issue #25.