Update all dependencies.
The eslint configuration now uses the new flat config.

Changes in eslint and typescript required a few changes to the code.
These should not change any behavior or functionality.

The "forkme" links have been removed from the test pages since the icon
that was used is no longer available and this is not the original
mathquill repository that the link was for anyway.