Turn on @typescript-eslint's strict type checking.
This requires a few changes, but mostly they are just typings which
don't change the actual generated code.  There several eslint-ignore
statements added for this that should be removed (with some effort).

Note that the pray and prayDirection methods has been removed.  Checking
the conditions directly in place requires about the same amount of code,
and has the benifit of revealing proper typings (not undefined mainly)
to typescript if the error is not thrown.