Dispense with the L, R enum.
Typescript continually has issues with that usage.  So now instead there
are "left" and "right" properties of nodes and the cursor.  The
direction type is just a type union of the strings 'left' and 'right'.

The advantages of a numeric value had already been taken away when this
project was initially converted to typescript.  For instance, you can't
multiply an enum value by -1 (to get the opposite direction).