Change focus/blur behavior to be more like a native HTML text input.
That is, when focus is lost within the window any selection is cleared.
When focus is regained the contents are all selected.  When the window
focus is lost the current state is preserved, and when focus returns it
is still in that state.

There is a new `preventBlur` option that is a method that will be passed
the blur event and the MathQuill api object.  That method should return
true if blurring the input should be prevented, and false if it should
be allowed.  Note that in this case the caller will need to ensure to
call the `blur` method of the MathQuill api object later when the input
should be blurred.  This is for PG's toolbar.  When focus leaves the
MathQuill input and goes to the toolbar, the MathQuill input remains
active.

The keyboard tab button no longer is used for navigation within an
editable MathQuill input.  It now is purely used for the usual HTML tab
order navigation.  The escape button still has the navigation behavior
it had before, and the primary navigation within an editable MathQuill
input is via the arrow keys (the same as for a native HTML input). The
`spaceBehavesLikeTab` option is now the `enableSpaceNavigation` option
and does the same thing it did before.