Improvements to focus/blur behavior with the preventBlur (now the blurWithCursor) option.
The focus handler still needs to run after a prevented blur, but a
select all should not occur.  So a more fine grained approach is needed.
If a blur is prevented, that needs to be saved and the focus handler
acts according to that saved state.

Also fix the 'select behaves correctly after blurring and re-focusing'
test.  That still doesn't work if the document is not visible.