Switch to using SVGs for roots and brackets.
This is also pulled from the Desmos code.  This looks nicer than trying
to use symbols scaled to the appropriate size.  That scaling was
gnerally not the best anyway.  This has the added advantage of not
needing to compute sizes.  Natural SVG scaling just works so much
better.

Also make the index a little bit larger and improve styling so that when
an nth root is selected the index is still visible.