Fix an issue with empty text blocks.
When focus is lost from the MathQuill input entirely and the cursor is
inside an empty text block, then the text block is not removed as it
should be.  Furthermore, if the text block is clicked on after that an
exception is thrown.  This ensures the text block is removed in this
case so that when focus returns the exception is not thrown.