Add `overflow: visible` to svgs used for brackets and such.
I am seeing clipping occur on the bottoms of brackets and braces in some
instances.  Adding `overflow: visible` prevents this.  I don't think
there is ever a situation where these svgs should be clipped, so I don't
think there is any downside to adding this.

At the default zoom level for me in both Firefox and Chrome the bottom
of a brace or bracket is clipped.  If I increase or decrease the zoom
level the clipping stops.  It also seems to usually only happen in the
first MathQuill input on the page, and not the later ones.  So overall
very odd.

Also run `npx update-browserslist-db@latest` because I am seeing
warnings about that.