Remove all of the default autoOperatorNames.
None of these are generally valid for answers in PG problem and having
them conflicts with units and string answers.  So now there are no
autoOperatorNames set by default (same as autoCommands).  They can still
be added if desired, and will continue to function as before if added.

The LaTeX commands are now created from the `BuiltInOpNames` constant
instead of constructing a new `Options` object and getting the default
`autoOperatorNames` from there.  That should never have been done that
way anyway, and was a hack that goes back to the upstream MathQuill
code.

This addresses https://github.com/openwebwork/webwork2/issues/2203 and
https://github.com/openwebwork/webwork2/issues/2681.