Update dependencies.
The `mocha` test suite has finally been updated to not use the
deprecated modules it was using.  So this fixes the warnings that were
being shown when running `npm ci` due that.

Fix the build with node 22.  The rollup libray build needed to be
changed since the `rollup-plugin-ts` package is deprecated and no longer
works with node 22.

Note that the `@awmottaz/prettier-plugin-void-html` now gives a warning
with node 20 about an unsupported engine.  However, everything still
works with node 20.

Also switch the output to es2022.

Finally, make this the official 0.11.0 release and remove the beta
subversion.