Fix the Symbola font glyphs that are now causing console log warnings.
There are three symbols in the font that are causing issues.  Those are
the names for the Greek letters Delta, Omega, and mu.  Those need to
specify that they are the Greek variants, and not the increment symbol,
ohm symbol, or micro symbol respectively.

The names were changed with fontforge.

To test this with webwork/PG copy the file `src/fonts/Symbola.woff2` in
this pull request to `/opt/webwork/pg/htdocs/node_modules/@openwebwork/mathquill/fonts`.
The console log warning will not appear with this version. The fonts
will also still be the same.  You can type the Greek letter Delta by
entering `\Delta`, for example.

I will bump the MathQuill version and publish a package containing this
fix at a later time.